+++ projects/bins/zocromas_duf/src/inc/duf_li_credel.h
@@ -2,6 +2,6 @@
 #  define DUF_LI_CREDEL_H
 
 duf_levinfo_t *duf_li_create( int count );
-duf_levinfo_t *duf_li_clone( const duf_levinfo_t * lisrc, size_t count );
+duf_levinfo_t *duf_li_clone( const duf_levinfo_t * lisrc, unsigned count );
 
 #endif
